Auto retry on 429 in merge
I was getting lots of rate limit messages while trying to merge a bunch
of stuff on my account.

I was getting responses such as:
x-ratelimit: {"name":"AUTHED_API_POST_LIMIT","period":1,"limit":1,
"remaining":0,"until":"2024-03-12T15:10:32Z"}
retry-after: 1

I added this retry logic which may be useful, but ultimately it
looks like the rate-limit tha tis set by the CLI should probably be used,
but is not currently used in the executor for the merge?